+++ addons/prestashop/synapxcookieconsent/synapxcookieconsent.php
@@ -53,7 +53,7 @@ class SynapxCookieConsent extends Module
     {
         $this->name = 'synapxcookieconsent';
         $this->tab = 'administration';
-        $this->version = '1.0.0';
+        $this->version = '1.0.1';
         $this->author = 'SynapxLab';
         $this->need_instance = 0;
         $this->bootstrap = true;