+++ src/js/helper.js
@@ -2,7 +2,7 @@
  * @synapxlab/cookie-consent - helper
  * Utilitaires pour le système de cookie consent
  * 
- * @version 2.5.0
+ * @version 2.5.1
  * @author SynapxLab <contact@synapxlab.com>
  * @license MIT
  */